requirements.txt

streamlit
Pillow
PyMuPDF
pandas
python-docx
gtts
python-pptx
